requirements.txt
Django==4.2.13
django-phonenumber-field==7.3.0
environs==11.0.0
phonenumbers==8.13.37
